delete node function

+++ src/application/ui/GGraph.java
@@ -6,6 +6,13 @@ import application.graph.Edge;
 import application.graph.Graph;
 import application.graph.Node;
 import application.ui.math.Vector2;
+import javafx.collections.FXCollections;
+import javafx.collections.ObservableList;
+import javafx.scene.control.TableColumn;
+import javafx.scene.control.TableView;
+import javafx.scene.control.TextField;
+import javafx.scene.control.TreeView;
+import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.layout.Pane;
 
 public class GGraph extends Graph{
@@ -15,12 +22,29 @@ public class GGraph extends Graph{
 	 */
 	private static final long serialVersionUID = 4967818067696204778L;
 	Pane drawPane;
+	TableView<GEdge> graphTable;
 	GNode selectedNode;
 	GEdge selectedEdge;
 	
-	public GGraph(Pane drawPane) {
+	public GGraph(Pane drawPane, TableView<GEdge> graphTable) {
 		super();
 		this.drawPane = drawPane;
+		
+		this.graphTable = graphTable;
+		TableColumn<GEdge, Integer> fromCol = new TableColumn<>("From");
+		fromCol.setPrefWidth(50);
+    	fromCol.setCellValueFactory(new PropertyValueFactory<>("from"));
+		TableColumn<GEdge, Integer> toCol = new TableColumn<>("To");
+		toCol.setPrefWidth(50);
+		toCol.setCellValueFactory(new PropertyValueFactory<>("to"));
+		TableColumn<GEdge, TextField> capCol = new TableColumn<>("Capacity");
+		capCol.setPrefWidth(120);
+		capCol.setCellValueFactory(new PropertyValueFactory<>("capField"));
+		TableColumn<GEdge, TextField> removeCol = new TableColumn<>("Remove");
+		removeCol.setPrefWidth(80);
+		removeCol.setCellValueFactory(new PropertyValueFactory<>("deleteBtn"));
+		graphTable.getColumns().addAll(fromCol, toCol, capCol, removeCol);
+		
 		selectedNode = null;
 		selectedEdge = null;
 	}
@@ -58,10 +82,22 @@ public class GGraph extends Graph{
 		adjList.add(new ArrayList<Integer>());
 	}
 	
-	public void deleteNode() throws Exception {
+	public boolean deleteNode() throws Exception {
+		if(selectedNode == null) return false;
+		for(Edge edge : edges) {
+			GEdge e = (GEdge)edge;
+			if(e.getFrom() == selectedNode.getId() || e.getTo() == selectedNode.getId())
+				e.remove();
+		}
 		super.deleteNode(selectedNode.getId());
+		for(int i=selectedNode.getId(); i<nodes.size(); i++) {
+			GNode n = (GNode)nodes.get(i);
+			n.updateLabel(Integer.valueOf(n.getId()).toString());
+		}
 		selectedNode.remove();
 		selectedNode = null;
+		updateGraphTable();
+		return true;
 	}
 	
 	private void addEdge(GNode n1, GNode n2) {
@@ -71,12 +107,22 @@ public class GGraph extends Graph{
 		}
 		edges.add(edge);
 		adjList.get(n1.getId()).add(n2.getId());
+		updateGraphTable();
 	}
 	
-	public void deleteEdge() throws Exception {
+	public boolean deleteEdge() throws Exception {
+		if(selectedEdge == null) return false;
 		super.deleteEdge(selectedEdge.getFrom(), selectedEdge.getTo());
 		selectedEdge.remove();
 		selectedEdge = null;
+		updateGraphTable();
+		return true;
+	}
+	
+	public void deleteEdge(GEdge e) throws Exception{
+		super.deleteEdge(e.getFrom(), e.getTo());
+		e.remove();
+		updateGraphTable();
 	}
 	
 	private boolean checkCornerClicked(int x, int y) {
@@ -101,12 +147,16 @@ public class GGraph extends Graph{
 							if(selectedEdge != e) { // another edge is selected
 								selectedEdge = ((GEdge)e);
 								selectedEdge.setSelected(true);
+								graphTable.getSelectionModel().clearSelection();
+								graphTable.getSelectionModel().select(idx);
 							}else { // the same edge is selected again
 								selectedEdge = null;
 							}
 						}else {  // newly selected edge
     						selectedEdge = (GEdge)e;
     						selectedEdge.setSelected(true);
+    						graphTable.getSelectionModel().clearSelection();
+    						graphTable.getSelectionModel().select(idx);
 						}
 					}else {	// newly created edge
 						addEdge(selectedNode, n);
@@ -139,6 +189,18 @@ public class GGraph extends Graph{
 	public GEdge getSelectedEdge() {
 		return selectedEdge;
 	}
+	
+	public void updateGraphTable() {
+		graphTable.getItems().clear();
+		ObservableList<GEdge> list = FXCollections.observableArrayList();
+		for(Edge e : edges) {
+			GEdge edge = (GEdge)e;
+			edge.capField.setText(Long.valueOf(e.getCapacity()).toString());
+			list.add(edge);
+		}
+		graphTable.setItems(list);
+		// graphTable.getSortOrder().add(graphTable.getColumns().get(0));
+	}
 
 	public void buildFromGraph(Graph g) {
 		nodeCount = 0;
@@ -165,12 +227,11 @@ public class GGraph extends Graph{
 		for(Node n : nodes) {
 			((GNode)n).remove();
 		}
-		
 		for(Edge e : edges) {
 			((GEdge)e).remove();
 		}
-		
 		super.clear();
+		updateGraphTable();
 	}
 	
 }
